requirements.txt
torch
fastai
torchvision
pandas 
numpy
scikit-learn
dvc
mlflow==2.7.1
notebook
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
gdown
autopep8
pytest
pytest-cov
codecov
mypy
dagshub
streamlit
boto3
python-dotenv
-e .
